requirements.txt
prompt_toolkit
alive-progress
openai
python-dotenv
